+++ src/features/fcConverter.ts
@@ -1,6 +1,9 @@
 import * as vscode from 'vscode';
 import * as child from 'child_process';
 
+import * as channel from '../other/outputChannel';
+import * as utils from '../other/utils';
+
 /*
 uses AnnoFCConverter from https://github.com/taubenangriff/AnnoFCConverter/
 */
@@ -12,14 +15,20 @@ export class FcConverter {
     const disposable = [
       vscode.commands.registerCommand('anno-modding-tools.convertFcCf7', (fileUri) => {
         if (fileUri) {
-          const res = child.execFileSync(converterPath, ['-y', '-r', fileUri.fsPath]);
-          console.log(res.toString());
+          const res = child.execFileSync(converterPath, [
+            '-r', fileUri.fsPath, 
+            '-o', utils.dontOverwrite(utils.swapExtension(fileUri.fsPath, '.fc', '.cf7'), '.cf7')
+          ]);
+          channel.log(res.toString());
         }
       }),
       vscode.commands.registerCommand('anno-modding-tools.convertCf7Fc', (fileUri) => {
         if (fileUri) {
-          const res = child.execFileSync(converterPath, ['-y', '-w', fileUri.fsPath]);
-          console.log(res.toString());
+          const res = child.execFileSync(converterPath, [
+            '-w', fileUri.fsPath,
+            '-o', utils.dontOverwrite(utils.swapExtension(fileUri.fsPath, '.cf7', '.fc'), '.fc')
+          ]);
+          channel.log(res.toString());
         }
       })
     ];