+++ src/features/rdpConverter.ts
@@ -8,17 +8,17 @@ export class RdpConverter {
     const disposable = [
       vscode.commands.registerCommand('anno-modding-tools.rdpToSimplified', async (fileUri) => {
         if (fileUri) {
-          await rdp.rdpToXml(fileUri.fsPath, path.dirname(fileUri.fsPath), true);
+          await rdp.rdpToXml(fileUri.fsPath, path.dirname(fileUri.fsPath), { simplify: true, dontOverwrite: true });
         }
       }),
       vscode.commands.registerCommand('anno-modding-tools.rdpToXml', async (fileUri) => {
         if (fileUri) {
-          await rdp.rdpToXml(fileUri.fsPath, path.dirname(fileUri.fsPath), false);
+          await rdp.rdpToXml(fileUri.fsPath, path.dirname(fileUri.fsPath), { simplify: false, dontOverwrite: true });
         }
       }),
       vscode.commands.registerCommand('anno-modding-tools.xmlToRdp', async (fileUri) => {
         if (fileUri) {
-          await rdp.xmlToRdp(fileUri.fsPath, path.dirname(fileUri.fsPath));
+          await rdp.xmlToRdp(fileUri.fsPath, path.dirname(fileUri.fsPath), { dontOverwrite: true });
         }
       })
     ];